feat: implement core billing engine and backend architecture with paise-based currency system

+++ backend/app/models/base.py
@@ -1,16 +1,16 @@
 # ============================================================
-# * PrimeSetu â€” Shoper9-Based Retail OS
-# * Zero Cloud Â. Sovereign Â. AI-Governed
+# * PrimeSetu — Shoper9-Based Retail OS
+# * Zero Cloud · Sovereign · AI-Governed
 # ============================================================
 # * System Architect   :  Jawahar R. M.
 # * Organisation       :  AITDL Network
 # * Project            :  PrimeSetu
-# * Â(c) 2026 â€” All Rights Reserved
+# * © 2026 — All Rights Reserved
 # * "Memory, Not Code."
 # ============================================================ #
 
 from __future__ import annotations
-from sqlalchemy import String, Integer, Numeric, Boolean, DateTime, text, ForeignKey, JSON
+from sqlalchemy import String, Integer, Numeric, Boolean, DateTime, text, ForeignKey, JSON, BigInteger
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from app.core.database import Base
 from datetime import datetime
@@ -54,11 +54,11 @@ class Product(Base):
     subcategory: Mapped[str] = mapped_column(String, nullable=True)
     size: Mapped[str] = mapped_column(String, nullable=True)
     color: Mapped[str] = mapped_column(String, nullable=True)
-    mrp: Mapped[float] = mapped_column(Numeric(15, 2), default=0.0)
-    wholesale_price: Mapped[float] = mapped_column(Numeric(15, 2), default=0.0)
-    staff_price: Mapped[float] = mapped_column(Numeric(15, 2), default=0.0)
-    cost_price: Mapped[float] = mapped_column(Numeric(15, 2), default=0.0)
-    tax_rate: Mapped[float] = mapped_column(Numeric(5, 2), default=18.0)
+    mrp: Mapped[int] = mapped_column(BigInteger, default=0)
+    wholesale_price: Mapped[int] = mapped_column(BigInteger, default=0)
+    staff_price: Mapped[int] = mapped_column(BigInteger, default=0)
+    cost_price: Mapped[int] = mapped_column(BigInteger, default=0)
+    tax_rate: Mapped[int] = mapped_column(Integer, default=18)
     is_tax_inclusive: Mapped[bool] = mapped_column(Boolean, default=True) # Shoper 9 Style
     is_inventory_item: Mapped[bool] = mapped_column(Boolean, default=True)
     is_active: Mapped[bool] = mapped_column(Boolean, default=True)
@@ -103,7 +103,7 @@ class Till(Base):
     code: Mapped[str] = mapped_column(String, unique=True)
     status: Mapped[str] = mapped_column(String, default="Closed") # Open, Closed, Locked, Idle
     current_cashier_id: Mapped[str] = mapped_column(String, nullable=True) # References auth.users or Partner
-    cash_collected: Mapped[float] = mapped_column(Numeric(15, 2), default=0.0)
+    cash_collected: Mapped[int] = mapped_column(BigInteger, default=0)
     last_opening_at: Mapped[datetime] = mapped_column(DateTime, nullable=True)
     last_closing_at: Mapped[datetime] = mapped_column(DateTime, nullable=True)
     created_at: Mapped[datetime] = mapped_column(DateTime, server_default=text("now()"))
@@ -122,16 +122,14 @@ class Transaction(Base):
     # [{"mode": "CASH", "amount": 1000}, {"mode": "CARD", "amount": 500, "ref": "123"}]
     payments: Mapped[Dict[str, Any]] = mapped_column(JSON, nullable=True)
     
-    subtotal: Mapped[float] = mapped_column(Numeric(15, 2), default=0.0)
-    discount_total: Mapped[float] = mapped_column(Numeric(15, 2), default=0.0)
-    tax_total: Mapped[float] = mapped_column(Numeric(15, 2), default=0.0)
-    net_payable: Mapped[float] = mapped_column(Numeric(15, 2), default=0.0)
+    subtotal: Mapped[int] = mapped_column(BigInteger, default=0)
+    discount_total: Mapped[int] = mapped_column(BigInteger, default=0)
+    tax_total: Mapped[int] = mapped_column(BigInteger, default=0)
+    net_payable: Mapped[int] = mapped_column(BigInteger, default=0)
     status: Mapped[str] = mapped_column(String, default="Finalized") # Finalized, Suspended, Void
     suspended_reason: Mapped[str] = mapped_column(String, nullable=True)
     created_at: Mapped[datetime] = mapped_column(DateTime, server_default=text("now()"))
     updated_at: Mapped[datetime] = mapped_column(DateTime, server_default=text("now()"), onupdate=text("now()"))
-
-
 
     items: Mapped[List["TransactionItem"]] = relationship("TransactionItem", back_populates="transaction", cascade="all, delete-orphan")
 
@@ -142,8 +140,8 @@ class CreditNote(Base):
     note_no: Mapped[str] = mapped_column(String, unique=True)
     customer_id: Mapped[uuid.UUID] = mapped_column(ForeignKey("customers.id"))
     original_transaction_id: Mapped[uuid.UUID] = mapped_column(ForeignKey("transactions.id"), nullable=True)
-    initial_amount: Mapped[float] = mapped_column(Numeric(15, 2))
-    balance_amount: Mapped[float] = mapped_column(Numeric(15, 2))
+    initial_amount: Mapped[int] = mapped_column(BigInteger)
+    balance_amount: Mapped[int] = mapped_column(BigInteger)
     expiry_date: Mapped[datetime] = mapped_column(DateTime, nullable=True)
     status: Mapped[str] = mapped_column(String, default="Active") # Active, Adjusted, Expired
     created_at: Mapped[datetime] = mapped_column(DateTime, server_default=text("now()"))
@@ -155,8 +153,8 @@ class AdvanceDeposit(Base):
     receipt_no: Mapped[str] = mapped_column(String, unique=True)
     customer_id: Mapped[uuid.UUID] = mapped_column(ForeignKey("customers.id"))
     purpose: Mapped[str] = mapped_column(String, nullable=True)
-    initial_amount: Mapped[float] = mapped_column(Numeric(15, 2))
-    balance_amount: Mapped[float] = mapped_column(Numeric(15, 2))
+    initial_amount: Mapped[int] = mapped_column(BigInteger)
+    balance_amount: Mapped[int] = mapped_column(BigInteger)
     status: Mapped[str] = mapped_column(String, default="Active") # Active, Adjusted, Refunded
     created_at: Mapped[datetime] = mapped_column(DateTime, server_default=text("now()"))
 
@@ -169,7 +167,7 @@ class PurchaseOrder(Base):
     store_id: Mapped[str] = mapped_column(ForeignKey("stores.id"))
     expected_delivery: Mapped[datetime] = mapped_column(DateTime, nullable=True)
     total_qty: Mapped[float] = mapped_column(Numeric(12, 3), default=0.0)
-    total_cost: Mapped[float] = mapped_column(Numeric(15, 2), default=0.0)
+    total_cost: Mapped[int] = mapped_column(BigInteger, default=0)
     status: Mapped[str] = mapped_column(String, default="Open") # Open, PartiallyReceived, Closed, Cancelled
     created_at: Mapped[datetime] = mapped_column(DateTime, server_default=text("now()"))
 
@@ -183,7 +181,7 @@ class PurchaseOrderItem(Base):
     product_id: Mapped[uuid.UUID] = mapped_column(ForeignKey("products.id"))
     qty: Mapped[float] = mapped_column(Numeric(12, 3))
     received_qty: Mapped[float] = mapped_column(Numeric(12, 3), default=0.0)
-    unit_cost: Mapped[float] = mapped_column(Numeric(15, 2))
+    unit_cost: Mapped[int] = mapped_column(BigInteger)
 
     purchase_order: Mapped["PurchaseOrder"] = relationship("PurchaseOrder", back_populates="items")
 
@@ -194,10 +192,10 @@ class TransactionItem(Base):
     transaction_id: Mapped[uuid.UUID] = mapped_column(ForeignKey("transactions.id", ondelete="CASCADE"))
     product_id: Mapped[uuid.UUID] = mapped_column(ForeignKey("products.id"))
     qty: Mapped[float] = mapped_column(Numeric(12, 3))
-    mrp: Mapped[float] = mapped_column(Numeric(15, 2))
-    discount_per: Mapped[float] = mapped_column(Numeric(5, 2), default=0.0)
-    tax_amount: Mapped[float] = mapped_column(Numeric(15, 2), default=0.0)
-    net_amount: Mapped[float] = mapped_column(Numeric(15, 2))
+    mrp: Mapped[int] = mapped_column(BigInteger)
+    discount_per: Mapped[int] = mapped_column(Integer, default=0)
+    tax_amount: Mapped[int] = mapped_column(BigInteger, default=0)
+    net_amount: Mapped[int] = mapped_column(BigInteger)
 
     transaction: Mapped["Transaction"] = relationship("Transaction", back_populates="items")
 
@@ -208,7 +206,7 @@ class SalesSlip(Base):
     slip_no: Mapped[str] = mapped_column(String, unique=True)
     store_id: Mapped[str] = mapped_column(ForeignKey("stores.id"))
     customer_mobile: Mapped[str] = mapped_column(String, nullable=True)
-    total_amount: Mapped[float] = mapped_column(Numeric(15, 2), default=0.0)
+    total_amount: Mapped[int] = mapped_column(BigInteger, default=0)
     is_active: Mapped[bool] = mapped_column(Boolean, default=True)
     is_converted: Mapped[bool] = mapped_column(Boolean, default=False)
     converted_transaction_id: Mapped[uuid.UUID] = mapped_column(ForeignKey("transactions.id"), nullable=True)
@@ -223,8 +221,8 @@ class SalesSlipItem(Base):
     slip_id: Mapped[uuid.UUID] = mapped_column(ForeignKey("sales_slips.id", ondelete="CASCADE"))
     product_id: Mapped[uuid.UUID] = mapped_column(ForeignKey("products.id"))
     qty: Mapped[float] = mapped_column(Numeric(12, 3))
-    mrp: Mapped[float] = mapped_column(Numeric(15, 2))
-    net_amount: Mapped[float] = mapped_column(Numeric(15, 2))
+    mrp: Mapped[int] = mapped_column(BigInteger)
+    net_amount: Mapped[int] = mapped_column(BigInteger)
 
     slip: Mapped["SalesSlip"] = relationship("SalesSlip", back_populates="items")
 
@@ -239,7 +237,7 @@ class Partner(Base):
     email: Mapped[str] = mapped_column(String, nullable=True)
     address: Mapped[str] = mapped_column(String, nullable=True)
     gst_no: Mapped[str] = mapped_column(String, nullable=True)
-    credit_limit: Mapped[float] = mapped_column(Numeric(15, 2), default=0.0)
+    credit_limit: Mapped[int] = mapped_column(BigInteger, default=0)
     loyalty_points: Mapped[int] = mapped_column(Integer, default=0)
     is_active: Mapped[bool] = mapped_column(Boolean, default=True)
     created_at: Mapped[datetime] = mapped_column(DateTime, server_default=text("now()"))
